+++ pymeasure/instruments/heidenhain/nd287.py
@@ -76,8 +76,8 @@ class ND287(Instrument):
     def id(self):
         """ String identification property for the device.
         """
-        self.adapter.connection.write("\x1BA0000")
-        id_str = self.adapter.connection.read_bytes(37).decode("utf-8")
+        self.write("\x1BA0000")
+        id_str = self.read_bytes(37).decode("utf-8")
         return id_str
 
     @property
@@ -101,9 +101,9 @@ class ND287(Instrument):
 
         :return: String with the error message as its contents.
         """
-        self.adapter.connection.write("\x1BA0301")
+        self.write("\x1BA0301")
         try:
-            err_str = self.adapter.connection.read_bytes(36).decode("utf-8")
+            err_str = self.read_bytes(36).decode("utf-8")
         except VisaIOError:
             err_str = None